Fixed key identification in ideal keyboard image

+++ contrast.hpp
@@ -1,11 +1,11 @@
 #ifndef CONTRAST_HPP
 #define CONTRAST_HPP
 
-#include<opencv2/highgui.hpp>
-#include<opencv/cv.hpp>
-#include<iostream>
-#include<stdlib.h>
-#include<math.h>
+#include <iostream>
+#include <math.h>
+#include <opencv/cv.hpp>
+#include <opencv2/highgui.hpp>
+#include <stdlib.h>
 
 using namespace std;
 using namespace cv;
@@ -14,9 +14,8 @@ using namespace cv;
 
 class Contrast {
 public:
-    void glg(Mat img);
-    void mergeLevels(Mat, float *, int*, float[], float[]);
-    void lookUpTable(float *, float [], float [], int*, float []);
-
+  void glg(Mat img);
+  void mergeLevels(Mat, double *, int *, float[], float[]);
+  void lookUpTable(float *, float[], float[], int *, float[]);
 };
 #endif
